evmctl: fix format-truncation warning in read_one_bank()
Building triggers:

warning: '%2.2d' directive writing between 2 and 11 bytes into a region
of size 4 [-Wformat-overflow=]

gcc computes the potential output width from i's type range rather
than its actual values; as a plain int it could be very large,
triggering a false-positive overflow warning. Narrow i to unsigned
short, which is more than enough for a PCR bank index, so gcc can
verify the write is safe.

+++ src/evmctl.c
@@ -2120,7 +2120,7 @@ static void extend_tpm_banks(struct template_entry *entry,
 static int read_one_bank(struct tpm_bank_info *tpm_bank, FILE *fp)
 {
 	char *p, pcr_str[8], buf[MAX_DIGEST_SIZE * 2 + 8];
-	int i = 0;
+	unsigned short i = 0;
 	int result = -1;
 	for (;;) {
 		p = fgets(buf, sizeof(buf), fp);
